requirements.txt
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
flatbuffers==24.3.25
idna==3.6
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.3
pandas==2.2.1
platformdirs==4.2.0
pyspark==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
six==1.16.0
tabulate==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
